fix(lang): resolve LangSpec table review findings
- src/search/strip.rs: gate extensionless dotfile stripping behind an
  extension check so files like .bashrc no longer get misclassified
- src/lang/spec.rs: drop the redundant default_defs_extract_name
  adapter, point extract_name at default_extract_definition_name
  directly
- src/lang/treesitter.rs / src/lang/go.rs / src/lang/mod.rs: hoist
  the QUERY_CACHE / with_query machinery into treesitter.rs so it's
  shared instead of duplicated
- src/search/siblings.rs: drop the now-duplicate
  StreamingIterator import and consume the hoisted with_query
