+++ src/search/siblings.rs
@@ -1,49 +1,8 @@
-use std::collections::HashMap;
-use std::sync::{LazyLock, Mutex};
-
 use streaming_iterator::StreamingIterator;
 
 use crate::lang::outline::outline_language;
+use crate::lang::treesitter::with_query;
 use crate::types::{Lang, OutlineEntry, OutlineKind};
-
-/// Global cache of compiled tree-sitter queries for sibling extraction.
-/// Keyed by `(node_kind_count, field_count, query_str_ptr)` so that distinct
-/// query strings for the same language (the main sibling query vs the Go
-/// receiver query) are stored under separate keys. We avoid `Language::name()`
-/// because ABI < 15 grammars (e.g. tree-sitter-kotlin-ng) return `None`.
-#[allow(clippy::type_complexity)]
-static QUERY_CACHE: LazyLock<Mutex<HashMap<(usize, usize, usize), tree_sitter::Query>>> =
-    LazyLock::new(|| Mutex::new(HashMap::new()));
-
-/// Look up or compile `query_str` for `ts_lang`, then invoke `f` with a
-/// reference to the cached `Query`.  Returns `None` if compilation fails.
-///
-/// `query_str` must be `'static` so its pointer address is stable across
-/// calls and can serve as part of the cache key.
-pub(crate) fn with_query<R>(
-    ts_lang: &tree_sitter::Language,
-    query_str: &'static str,
-    f: impl FnOnce(&tree_sitter::Query) -> R,
-) -> Option<R> {
-    use std::collections::hash_map::Entry;
-    // Pointer address distinguishes different queries for the same language.
-    let key = (
-        ts_lang.node_kind_count(),
-        ts_lang.field_count(),
-        query_str.as_ptr() as usize,
-    );
-    let mut cache = QUERY_CACHE
-        .lock()
-        .unwrap_or_else(std::sync::PoisonError::into_inner);
-    let query = match cache.entry(key) {
-        Entry::Occupied(e) => e.into_mut(),
-        Entry::Vacant(e) => {
-            let q = tree_sitter::Query::new(ts_lang, query_str).ok()?;
-            e.insert(q)
-        }
-    };
-    Some(f(query))
-}
 
 /// A sibling field or method resolved from the same parent struct/class/impl.
 #[derive(Debug)]
