+++ src/search/strip.rs
@@ -18,9 +18,11 @@ use crate::types::FileType;
 /// non-code), so they are mapped to the JS/TS family here to preserve the
 /// historical strip behavior for those files.
 fn detect_lang(path: &Path) -> Option<StripFamily> {
-    if let FileType::Code(lang) = detect_file_type(path) {
-        if let Some(family) = spec(lang).strip_family {
-            return Some(family);
+    if path.extension().is_some() {
+        if let FileType::Code(lang) = detect_file_type(path) {
+            if let Some(family) = spec(lang).strip_family {
+                return Some(family);
+            }
         }
     }
     match path.extension()?.to_str()? {
@@ -295,6 +297,13 @@ mod tests {
     fn ruby_not_supported() {
         let content = "def foo\n  puts 'hi'\nend\n";
         let skip = strip_noise(content, &path("rb"), Some((1, 3)));
+        assert!(skip.is_empty());
+    }
+
+    #[test]
+    fn extensionless_bash_dotfile_not_stripped() {
+        let content = "# just a comment\necho hi\n";
+        let skip = strip_noise(content, &PathBuf::from(".bashrc"), Some((1, 2)));
         assert!(skip.is_empty());
     }
